# internal/docs/swagger.yaml
basePath: /api
definitions:
  category.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  category.CreateCategoryRequest:
    properties:
      name:
        maxLength: 50
        type: string
    required:
    - name
    type: object
  category.UpdateCategoryRequest:
    properties:
      name:
        maxLength: 50
        type: string
    type: object
  item.CreateItemRequest:
    properties:
      category_id:
        type: integer
      expiry_date:
        example: "2025-12-31"
        type: string
      name:
        maxLength: 100
        type: string
      notes:
        type: string
      quantity:
        minimum: 0
        type: number
      unit:
        maxLength: 20
        type: string
    required:
    - name
    - quantity
    - unit
    type: object
  item.Item:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      expiry_date:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      quantity:
        type: number
      unit:
        type: string
      updated_at:
        type: string
    type: object
  item.UpdateItemRequest:
    properties:
      category_id:
        type: integer
      expiry_date:
        type: string
      name:
        maxLength: 100
        type: string
      notes:
        type: string
      quantity:
        minimum: 0
        type: number
      unit:
        maxLength: 20
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple inventory tracking REST API built with Gin.
  title: Fridge API
  version: "1.0"
paths:
  /categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/category.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a category
      parameters:
      - description: Category to create
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/category.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Deleet a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a category
      tags:
      - categories
    get:
      description: Retrieve a single category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update the name of a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category fields to update
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/category.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a category
      tags:
      - categories
  /items:
    get:
      description: Retrieve a list of all items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/item.Item'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create an item
      parameters:
      - description: Item to create
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/item.CreateItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/item.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new item
      tags:
      - items
  /items/{id}:
    delete:
      description: Delete an item by its ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an item
      tags:
      - items
    get:
      description: Retrieve a single item by its ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an item by ID
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Update the attributes of an item by ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item fields to update
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/item.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an item
      tags:
      - items
swagger: "2.0"